refactor(database): create postgresql package
